requirements.txt
flask~=3.0.3
websocket-client~=1.8.0
python-dotenv~=1.0.1
nanoid~=2.0.0
requests~=2.32.3
pyjwt~=2.9.0
werkzeug~=3.0.6
openai-whisper~=20240930
sqlalchemy~=2.0.40
bs4~=0.0.2
beautifulsoup4~=4.13.4
urllib3~=1.26.6
pyttsx3~=2.98
pymysql~=1.1.1
flask-cors~=5.0.0
flask-migrate~=4.1.0
celery~=5.5.2
redis~=5.2.1
alembic~=1.14.1